# `os.isCygwinPty`: Fix a bug, replace kernel32 call, and optimize